+++ pint2vela/tests/test_data_files.py
@@ -14,6 +14,9 @@ datasets = [
     "sim_sw",
     "sim_fdjumpdm",
     "sim_cm",
+    "sim3",
+    "sim4",
+    # "sim6",
 ]